Moved JsonError and ErrorDetail to new package
