+++ jaxrs/src/test/scala/com/cerner/beadledom/jaxrs/exceptionmapper/ThrowableExceptionMapperSpec.scala
@@ -1,6 +1,6 @@
 package com.cerner.beadledom.jaxrs.exceptionmapper
 
-import com.cerner.beadledom.jaxrs.models.JsonError
+import com.cerner.beadledom.json.common.models.JsonError
 import com.cerner.beadledom.jaxrs.provider.{FakeRepository, FakeResource}
 
 import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider
