+++ json-common/README.rst
@@ -1,0 +1,6 @@
+.. _beadledom-json-common:
+
+beadledom-json-common
+=====================
+
+`Check out our docs for more information <http://cerner.github.io/beadledom>`_
